More checks for interruption along GenvisisPipeline;

+++ src/gwas/Qc.java
@@ -2,6 +2,7 @@ package gwas;
 
 import java.io.*;
 import java.util.*;
+
 import common.*;
 
 public class Qc {
@@ -32,30 +33,37 @@ public class Qc {
 			log.report(ext.getTime() + "]\tRunning --freq");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --freq --out freq --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"markerQC/missing.imiss")) {
 			log.report(ext.getTime() + "]\tRunning --missing");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --missing --out missing --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"markerQC/test.missing.missing")) {
 			log.report(ext.getTime() + "]\tRunning --test-missing");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --test-missing --out test.missing --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"markerQC/hardy.hwe")) {
 			log.report(ext.getTime() + "]\tRunning --hardy");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --hardy --out hardy --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"markerQC/mishap.missing.hap")) {
 			log.report(ext.getTime() + "]\tRunning --test-mishap");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --test-mishap --out mishap --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"markerQC/gender.assoc")) {
 			log.report(ext.getTime() + "]\tRunning --assoc gender");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --pheno plink.fam --mpheno 3 --assoc --out gender --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"markerQC/gender.missing")) {
 			log.report(ext.getTime() + "]\tRunning --test-missing gender");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --pheno plink.fam --mpheno 3 --test-missing --out gender --noweb", dir+"markerQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		
 		if (!Files.exists(dir+"markerQC/miss_drops.dat")) {
 			new File(dir+"markerQC/miss.crf").delete();
@@ -69,54 +77,65 @@ public class Qc {
 			    // ERROR  TODO not sure if we should quit here; for now, continue;
 			}
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 
 		new File(dir+"sampleQC/").mkdirs();
 		if (!Files.exists(dir+"sampleQC/plink.bed")) {
 			log.report(ext.getTime() + "]\tRunning --exclude miss_drops.dat");
 			CmdLine.runDefaults("plink --bfile ../plink --exclude ../markerQC/miss_drops.dat --make-bed --noweb", dir+"sampleQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"sampleQC/missing.imiss")) {
 			log.report(ext.getTime() + "]\tRunning --missing");
 			CmdLine.runDefaults("plink --bfile plink --maf 0 --geno 1 --mind 1 --missing --out missing --noweb", dir+"sampleQC/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		
 		new File(dir+"ldPruning/").mkdirs();
 		if (!Files.exists(dir+"ldPruning/plink.bed")) {
 			log.report(ext.getTime() + "]\tRunning --mind 0.05 (removes samples with callrate <95% for the markers that did pass QC)");
 			CmdLine.runDefaults("plink --bfile ../sampleQC/plink --mind 0.05 --make-bed --noweb", dir+"ldPruning/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"ldPruning/plink.prune.in")) {
 			log.report(ext.getTime() + "]\tRunning --indep-pairwise 50 5 0.3");
 			CmdLine.runDefaults("plink --noweb --bfile plink --indep-pairwise 50 5 0.3", dir+"ldPruning/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 
 		new File(dir+"genome/").mkdirs();
 		if (!Files.exists(dir+"genome/plink.bed")) {
 			log.report(ext.getTime() + "]\tRunning --extract plink.prune.in");
 			CmdLine.runDefaults("plink --bfile ../ldPruning/plink --extract ../ldPruning/plink.prune.in --make-bed --noweb", dir+"genome/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"genome/plink.genome")) {
 			log.report(ext.getTime() + "]\tRunning --genome"+(keepGenomeInfoForRelatedsOnly?" --min 0.1":""));
 			CmdLine.runDefaults("plink --noweb --bfile plink --genome"+(keepGenomeInfoForRelatedsOnly?" --min 0.1":""), dir+"genome/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!keepGenomeInfoForRelatedsOnly && !Files.exists(dir + "genome/mds20.mds")) {
 			log.report(ext.getTime() + "]\tRunning --mds-plot 20");
 			CmdLine.runDefaults("plink --bfile plink --read-genome plink.genome --cluster --mds-plot 20 --out mds20 --noweb", dir+"genome/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"genome/plink.genome_keep.dat")) {
 			log.report(ext.getTime() + "]\tRunning flagRelateds");
 			Plink.flagRelateds(dir+"genome/plink.genome", dir+"genome/plink.fam", dir+"markerQC/missing.imiss", dir+"genome/lrr_sd.xln", Plink.FLAGS, Plink.THRESHOLDS, 4, false);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		
 		new File(dir+"ancestry/").mkdirs();
 		if (!Files.exists(dir+"ancestry/unrelateds.txt")) {
 			log.report(ext.getTime() + "]\tCopying genome/plink.genome_keep.dat to ancestry/unrelateds.txt");
 			Files.copyFile(dir+"genome/plink.genome_keep.dat", dir+"ancestry/unrelateds.txt");
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		if (!Files.exists(dir+"ancestry/plink.bed")) {
 			log.report(ext.getTime() + "]\tRunning --extract plink.prune.in (again, this time to ancestry/)");
 			CmdLine.runDefaults("plink --bfile ../genome/plink --make-bed --noweb", dir+"ancestry/", log);
 		}
+        if (Thread.currentThread().isInterrupted()) { throw new RuntimeException(new InterruptedException()); }
 		
 		ancestry(dir+"ancestry/");
